Create gpupdate.vbs
This requires you to have PSTools downloaded and extracted to C:\TOOLS\ for this to work as is. The script should be run on a DC and as a Domain Admin. This will loop through AD and create a separate instance for each computer and run GPUpdate /force on the system.

PSTools Download: http://technet.microsoft.com/en-us/sysinternals/bb896649.aspx